Make saturating uint8_t narrowings explicit
- avr-libc's UINT8_MAX is unsigned int, so `cond ? UINT8_MAX : x` narrows to uint8_t; Alpine's avr-gcc 15.2 flagged it under -Wconversion (`asdf_ring_drop`, `asdf_process_r`, `asdf_update_r`)

+++ src/asdf.c
@@ -225,7 +225,7 @@ void asdf_tick_r(asdf_t *kb, uint8_t elapsed_ms) {
 // COMPLEXITY: 4
 //
 void asdf_process_r(asdf_t *kb, uint16_t elapsed_ms) {
-    uint8_t elapsed = (elapsed_ms > UINT8_MAX) ? UINT8_MAX : (uint8_t)elapsed_ms;
+    uint8_t elapsed = (elapsed_ms > UINT8_MAX) ? (uint8_t)UINT8_MAX : (uint8_t)elapsed_ms;
 
     if (!elapsed) {
         return;
@@ -258,7 +258,7 @@ void asdf_process_r(asdf_t *kb, uint16_t elapsed_ms) {
 // COMPLEXITY: 2
 //
 void asdf_update_r(asdf_t *kb, uint16_t elapsed_ms) {
-    uint8_t elapsed = (elapsed_ms > UINT8_MAX) ? UINT8_MAX : (uint8_t)elapsed_ms;
+    uint8_t elapsed = (elapsed_ms > UINT8_MAX) ? (uint8_t)UINT8_MAX : (uint8_t)elapsed_ms;
 
     if (elapsed) {
         asdf_tick_r(kb, elapsed);
